Refactor admin clients into modular components
Split the All Clients admin page into reusable components and helpers: ClientsStats, ClientsToolbar, ClientsTable, ClientViewModal, ClientImpersonateModal, and clientsHelpers. Extracted inline CSS into src/styles/admin/AdminAllClients.css and updated AdminAllClients.jsx to import and use the new components. Also cleaned up API call formatting, state spacing, and minor UI logic (loading/empty states delegated to ClientsTable) while preserving existing behavior for toggling status and impersonation flows.
